+++ web/src/web/pages/home/__init__.py
@@ -1,0 +1,18 @@
+CAMPUSES = [
+    {
+        "value": "UNIVERSITY COLLEGE HOSPITAL CAMPUS",
+        "label": "UCH",
+        "default_dept": "UCH T03 INTENSIVE CARE",
+    },
+    {
+        "value": "GRAFTON WAY BUILDING",
+        "label": "GWB",
+        "default_dept": "GWB L01 CRITICAL CARE",
+    },
+    {
+        "value": "WESTMORELAND STREET",
+        "label": "WMS",
+        "default_dept": "WMS W01 CRITICAL CARE",
+    },
+    {"value": "QUEEN SQUARE CAMPUS", "label": "NHNN", "default_dept": "NHNN C1 NCCU"},
+]
