+++ web/src/web/pages/home/cyto_style_sheet.json
