--- web/src/web/pages/sitrep/__init__.py
@@ -1,18 +1,0 @@
-CAMPUSES = [
-    {
-        "value": "GRAFTON WAY BUILDING",
-        "label": "GWB",
-        "default_dept": "GWB L01 CRITICAL CARE",
-    },
-    {"value": "QUEEN SQUARE CAMPUS", "label": "NHNN", "default_dept": "NHNN C1 NCCU"},
-    {
-        "value": "UNIVERSITY COLLEGE HOSPITAL CAMPUS",
-        "label": "UCH",
-        "default_dept": "UCH T03 INTENSIVE CARE",
-    },
-    {
-        "value": "WESTMORELAND STREET",
-        "label": "WMS",
-        "default_dept": "WMS W01 CRITICAL CARE",
-    },
-]
